# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Silly Walk Grant Application Orchestrator
  description: |-
    Backend service for managing silly walk grant applications for the Ministry of Silly Walks.
    
    This API allows for the submission, retrieval, and listing of silly walk grant applications.
    Each application is validated, scored for silliness, and stored securely.
  version: 1.0.0
  contact:
    email: ministry@sillywalks.gov.example
servers:
  - url: /api/v1
    description: API Version 1
tags:
  - name: Applications
    description: Operations related to silly walk grant applications
paths:
  /applications:
    post:
      tags:
        - Applications
      summary: Submit a new silly walk grant application
      description: |-
        Accepts a new silly walk grant application, validates input, calculates
        silliness score, persists the application, and returns the application ID and score.
      operationId: submitApplication
      security:
        - ApiKeyAuth: []
      requestBody:
        description: Application data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationSubmissionRequest'
      responses:
        '201':
          description: Application submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSubmissionResponse'
        '400':
          description: Invalid application data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Applications
      summary: List all silly walk grant applications
      description: |-
        Retrieves a list of all silly walk grant applications with pagination support.
      operationId: listApplications
      parameters:
        - name: skip
          in: query
          description: Number of applications to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Maximum number of applications to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Applications retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationListResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /applications/{application_id}:
    get:
      tags:
        - Applications
      summary: Get a silly walk grant application by ID
      description: |-
        Retrieves detailed information about a specific silly walk grant application.
      operationId: getApplicationById
      parameters:
        - name: application_id
          in: path
          description: ID of the application to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          description: Application not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ApplicationSubmissionRequest:
      type: object
      properties:
        applicant_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the applicant
          example: John Cleese
        walk_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Name of the silly walk
          example: The Ministry Shuffle
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Detailed description of the silly walk
          example: A walk involving high knee lifts, occasional hopping, and precisely timed briefcase movements.
        has_briefcase:
          type: boolean
          description: Whether the walk involves a briefcase
          example: true
        involves_hopping:
          type: boolean
          description: Whether the walk involves hopping
          example: true
        number_of_twirls:
          type: integer
          minimum: 0
          maximum: 100
          description: Number of twirls in the walk
          example: 5
      required:
        - applicant_name
        - walk_name
        - description
        - has_briefcase
        - involves_hopping
        - number_of_twirls
    ApplicationSubmissionResponse:
      type: object
      properties:
        application_id:
          type: string
          format: uuid
          description: Unique identifier for the application
          example: 123e4567-e89b-12d3-a456-426614174000
        silliness_score:
          type: integer
          description: Calculated silliness score
          example: 32
        status:
          type: string
          description: Current status of the application
          example: PendingReview
      required:
        - application_id
        - silliness_score
        - status
    ApplicationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the application
          example: 123e4567-e89b-12d3-a456-426614174000
        applicant_name:
          type: string
          description: Name of the applicant
          example: John Cleese
        walk_name:
          type: string
          description: Name of the silly walk
          example: The Ministry Shuffle
        description:
          type: string
          description: Detailed description of the silly walk
          example: A walk involving high knee lifts, occasional hopping, and precisely timed briefcase movements.
        has_briefcase:
          type: boolean
          description: Whether the walk involves a briefcase
          example: true
        involves_hopping:
          type: boolean
          description: Whether the walk involves hopping
          example: true
        number_of_twirls:
          type: integer
          description: Number of twirls in the walk
          example: 5
        silliness_score:
          type: integer
          description: Calculated silliness score
          example: 32
        status:
          type: string
          description: Current status of the application
          example: PendingReview
        submission_timestamp:
          type: string
          format: date-time
          description: When the application was submitted
          example: 2023-05-21T14:30:00Z
      required:
        - id
        - applicant_name
        - walk_name
        - description
        - has_briefcase
        - involves_hopping
        - number_of_twirls
        - silliness_score
        - status
        - submission_timestamp
    ApplicationListResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponse'
          description: List of applications
        total:
          type: integer
          description: Total number of applications
          example: 42
      required:
        - applications
        - total
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: Invalid application data.
      required:
        - detail